Guidebook tweaks
Fix a typo in the text for encumbrance in the Status section.

Fix a typo in Guidebook.mn for non-fatal status condtions that
resulted in bold font for the entire Stun/Conf/Hallu line.

For Guidebook.ps, force
 Status
     Hunger:
onto two lines instead of being bunched up together on one line.
(For Guidebook.txt, it's already two lines.)

For both Guidebook.txt and Guidebook.ps, force some extra vertical
separation between the set of one-letter commands and the set of
extended commands, between the set of extended commands and the
set of Meta-key commands, and between the set of Meta-key commands
and the extra one-letter commands available when number_pad is on.
I don't know whether Guidebook.tex's output would look better with
something similar.