more pluralization bits
Handle more *man and *men cases.

Some plural usage of completely made up fruit names that should
be entered in singular form but have what appears to be a
valid plural name it will end up singularized. Not much
we can do about that for ficticious words.

For instance, if you try to name your fruit bigmen or snakemen
and you intended that to be the singular name, NetHack will likely
singularize it to bigman or snakeman.

Many real dictionary words that end in "men", however, should
be handled a wee bit better now. A real word such as stamen,
for example.